Add global root node (#388)
Summary

- Add migration to add global root node and admin user which owns it
- On data load, pull all public objects
- In graph store, ensure there's a global root node and relation b/w
user and that node. If during instantiation we do create a global root
node locally, we use the same id as it has in the db, and don't push the
change. So this is just a local copy that we've hard-coded into the app.
- The default home page now shows the global root node in the
breadcrumbs

Completes ENT-3940

## How to test

- Create/use a test db
- Checkout `taylor/global-root` and do `yarn db:reset` 
- Run `yarn dev` and open the app. 
- You should see a "Global Root" node in the breadcrumbs above the user
node.
- You should see a sublist relation connecting the global root to the
user node.
- If you look at the db, you should see a single global root node

+++ src/db/graphNodes.ts
@@ -3,6 +3,7 @@ import { and, eq } from "drizzle-orm";
 import { SerializedNode } from "@/app/persistence/SerializedData";
 import { graphNodeTable, relationListsTable } from "@/db/schema";
 import { MewDbTransaction } from "@/db/types";
+import { GLOBAL_ROOT_ID, USER_ROOT_ID } from "@/lib/constants";
 
 export const createNode = async (tx: MewDbTransaction, node: SerializedNode) => {
   await tx.insert(graphNodeTable).values({
@@ -47,6 +48,13 @@ export const updateNode = async (tx: MewDbTransaction, oldProps: SerializedNode,
 };
 
 export const deleteNode = async (tx: MewDbTransaction, node: SerializedNode) => {
+  if (node.id === USER_ROOT_ID) {
+    throw new Error("Cannot delete user root node");
+  }
+  if (node.id === GLOBAL_ROOT_ID) {
+    throw new Error("Cannot delete global root node");
+  }
+
   // Delete all relationLists entries that reference this node
   await tx.delete(relationListsTable).where(eq(relationListsTable.nodeId, node.id));
